Moving to a modular structure for the benchmarks
I think it would be cool to have each set of conditions in it's own little folder with the protocol named protocol.xml, the output files all in a folder called out/, etc. so that it's easy to go back and compare the different sets of conditions against each other. For now, I'm sticking to enzdes out which will be in out/score.sc